Backend/app/requirements.txt
langchain
langchain-community
langchain-huggingface
chromadb
python-dotenv
fastapi
uvicorn
transformers
accelerate
sentence-transformers
torch
bitsandbytes
tiktoken